+++ src/command_modules/azure-cli-rdbms/azure/cli/command_modules/rdbms/_params.py
@@ -101,16 +101,8 @@ def load_arguments(self, _):    # pylint: disable=too-many-statements
             c.argument('virtual_network_subnet_id', options_list=['--subnet'], help='Name or ID of the subnet that allows access to an Azure Postgres Server. If subnet name is provided, --vnet-name must be provided.')
             c.argument('ignore_missing_vnet_service_endpoint', options_list=['--ignore-missing-endpoint', '-i'], help='Create vnet rule before virtual network has vnet service endpoint enabled', arg_type=get_three_state_flag())
 
-        with self.argument_context('postgres server vnet-rule create') as c:
-            c.extra('vnet_name', options_list=['--vnet-name'], help='The virtual network name', validator=validate_subnet)
-
-        with self.argument_context('postgres server vnet-rule update') as c:
-            c.extra('vnet_name', options_list=['--vnet-name'], help='The virtual network name', validator=validate_subnet)
-
-        with self.argument_context('mysql server vnet-rule create') as c:
-            c.extra('vnet_name', options_list=['--vnet-name'], help='The virtual network name', validator=validate_subnet)
-
-        with self.argument_context('mysql server vnet-rule update') as c:
+    for scope in ['postgres server vnet-rule create', 'postgres server vnet-rule update', 'mysql server vnet-rule create', 'mysql server vnet-rule update']:
+        with self.argument_context(scope) as c:
             c.extra('vnet_name', options_list=['--vnet-name'], help='The virtual network name', validator=validate_subnet)
 
     for scope in ['mysql server configuration', 'postgres server configuration']: